# documents/openapi.yaml
openapi: 3.0.0
info:
  title: Bento API
  description: API documentation for the Bento site user management system
  version: 1.0.0
servers:
  - url: https://stg.bento-o.site
    description: Secure server (HTTPS)
  - url: http://stg.bento-o.site
    description: Insecure server (HTTP for local testing)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # Optional, but clarifies the token type.
paths:
...
  /notes:
    get:
      tags:
        - Notes
      summary: Get all notes
      description: Retrieves all notes in the system.
      operationId: getNotes
      responses:
        200:
          description: A list of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    note_id:
                      type: integer
                      example: 1
                    title:
                      type: string
                      example: "My first note"
                    folder:
                      type: string
                      example: "deep learning"
                    created_at:
                      type: string
                      format: date-time
                      example: "2021-10-01 12:00:00"
                    duration:
                      type: string
                      example: "01:00:00"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The provided token is invalid. Please check and provide a valid token."
    post:
      tags:
        - Notes
      summary: Create a new note
      description: Creates a new note in the system.
      operationId: postNote
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "test title"
                folder:
                  type: string
                  example: "deep-learning"
                file:
                  type: string
                  format: binary
                  description: The file to be uploaded.
                bookmarks:
                  type: array
                  items:
                    type: object
                  description: Optional list of bookmarks in JSON format.
                  example: [{"timestamp":"00:22:33"},{"timestamp":"00:22:34"}]
                  nullable: true
                memos:
                  type: array
                  items:
                    type: object
                  description: Optional list of memos in JSON format.
                  example: [{"timestamp":"00:22:33","text":"메모1"},{"timestamp":"00:22:34","text":"메모2"}]
                  nullable: true
      responses:
        201:
          description: Note created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Note created successfully"
                  id:
                    type: integer
                    example: 1
        400:
          description: Validation failed for the input data
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: object
                    additionalProperties:
                      type: string
                example:
                  title: "Title is mandatory"
                  folder: "Folder is mandatory"
  /notes/search:
    get:
      tags:
        - Notes
      summary: Search notes
      description: Searches notes in the system based on the provided query, date range, or other criteria.
      operationId: searchNotes
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
          description: Text to search for within the notes' content or title.
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
            example: "2024-12-01"
          description: Start date of the search range in `yyyy-MM-dd` format.
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
            example: "2024-12-03"
          description: End date of the search range in `yyyy-MM-dd` format.
      responses:
        200:
          description: A list of notes matching the search criteria.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    noteId:
                      type: integer
                      example: 1
                    title:
                      type: string
                      example: "My first note"
                    folder:
                      type: string
                      example: "deep learning"
                    createdAt:
                      type: string
                      format: date-time
                      example: "2024-12-03 03:19:32"
                    duration:
                      type: string
                      example: "01:00:00"
                    matches:
                      type: array
                      items:
                        type: object
                        properties:
                          text:
                            type: string
                            example: "Example matching text"
                          timestamp:
                            type: string
                            format: time
                            example: "00:00:41"
        400:
          description: Bad Request - Invalid parameters or date formats.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The provided date format is invalid. Please provide a valid date in the format yyyy-MM-dd."
        401:
          description: Unauthorized - Invalid authentication token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "The provided token is invalid. Please check and provide a valid token."
  /notes/{noteId}:
    get:
      tags:
        - Notes
      summary: Get note details
      description: Retrieves details of a specific note by its ID.
      security:
        - bearerAuth: []
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Note details
          content:
            application/json:
              schema:
                type: object
                properties:
                  noteId:
                    type: integer
                    example: 1
                  title:
                    type: string
                    example: "My Note"
                  folder:
                    type: string
                    example: "My Folder"
                  createdAt:
                    type: string
                    format: date-time
                    example: "2021-10-01 12:00:00"
                  duration:
                    type: string
                    example: "01:00:00"
                  content:
                    type: object
                    properties:
                      script:
                        type: array
                        items:
                          type: object
                          properties:
                            memo:
                              type: string
                              example: "메모 내용"
                            bookmark:
                              type: boolean
                              example: true
                            text:
                              type: string
                              example: "네 그래서 저희 진행 상황 공유드리고 그다음에 금주 업무 계획 같이 살펴보실 예정이고요."
                            speaker:
                              type: string
                              example: "Speaker 1"
                            timestamp:
                              type: string
                              example: "00:00:00"
                      speaker:
                        type: array
                        items:
                          type: string
                        example:
                          - "Speaker 1"
                          - "Speaker 2"
                  topics:
                    type: array
                    items:
                      type: string
                    example:
                      - "Deep Learning"
                      - "Machine Learning"
                  bookmarks:
                    type: array
                    items:
                      type: object
                      properties:
                        timestamp:
                          type: string
                          example: "00:00:00"
                  memos:
                    type: array
                    items:
                      type: object
                      properties:
                        text:
                          type: string
                          example: "This is a memo"
                        timestamp:
                          type: string
                          example: "00:00:00"
                  ai:
                    type: array
                    items:
                      type: string
                    example:
                      - "This is a summary of the note."

        404:
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: object
                    additionalProperties:
                      type: string
                example:
                  message: "Note not found"
    delete:
      tags:
        - Notes
      summary: Delete a note
      description: Deletes a specific note by its ID.
      security:
        - bearerAuth: []
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Note deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Note deleted successfully"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: object
                    additionalProperties:
                      type: string
                example:
                  message: "Note not found"
    patch:
      tags:
        - Notes
      summary: Update a note
      description: Updates the details of a specific note by its ID.
      security:
        - bearerAuth: []
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "New Title"
                folderName:
                  type: string
                  example: "new folder"
      responses:
        200:
          description: Note updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Note updated successfully"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: object
                    additionalProperties:
                      type: string
                example:
                  message: "Note not found"
  /notes/{noteId}/status:
    patch:
      tags:
        - Notes
      summary: Update note status
      description: Updates the status of a specific note by its ID.
      security:
        - bearerAuth: []
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [ "PROCESSING", "COMPLETE", "FAILED" ]
                  example: "COMPLETE"
      responses:
        200:
          description: Note status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Note status updated successfully"
        400:
          description: Validation failed for the input data
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: object
                    additionalProperties:
                      type: string
                example:
                  status: "Invalid status provided. Please provide a valid status."
  /notes/folders:
    get:
      tags:
        - Notes
      summary: Get all folders
      description: Retrieves a list of folders for the authenticated user.
      security:
        - bearerAuth: []
      responses:
        200:
          description: A list of folders
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    folder_id:
                      type: integer
                      example: 1
                    folderName:
                      type: string
                      example: "deep learning"
    post:
      tags:
        - Notes
      summary: Create a new folder
      description: Creates a new folder in the system.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                folderName:
                  type: string
                  example: "New Folder name"
      responses:
        201:
          description: Note created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Folder created successfully"
                  id:
                    type: integer
                    example: 1
        400:
          description: Validation failed for the input data
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: object
                    additionalProperties:
                      type: string
                example:
                  message: "Folder already exists || Folder name is required"
  /notes/folders/{folderId}:
    get:
      tags:
        - Notes
      summary: Get notes by folder
      description: Retrieves notes within a specific folder for the authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - name: folderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: A list of notes in the folder
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    note_id:
                      type: integer
                      example: 1
                    title:
                      type: string
                      example: "My first note"
                    folder:
                      type: string
                      example: "deep learning"
                    created_at:
                      type: string
                      format: date-time
                      example: "2021-10-01 12:00:00"
                    duration:
                      type: string
                      example: "01:00:00"
  /notes/{noteId}/ai-summary:
    get:
      tags:
        - Notes
      summary: Get AI-generated summary
      description: Retrieves an AI-generated summary for the specified note.
      security:
        - bearerAuth: []
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: AI summary retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: string
                    example: "This is a summary of the note."
  /notes/download/{fileName}:
    get:
      tags:
        - Notes
      summary: Download a file
      description: (테스트용 API) Downloads a file associated with a note by its filename.
      security:
        - bearerAuth: []
      parameters:
        - name: fileName
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: File downloaded successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        400:
          description: Validation failed for the input data
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: object
                    additionalProperties:
                      type: string
                example:
                  message: "File not found or unreadable"
...
  /bookmarks/note/{noteId}:
    get:
      tags:
        - Bookmarks
      summary: Get bookmarks by note ID
      description: Retrieves bookmarks for a specific note.
      operationId: getBookmarksByNoteId
      security:
        - bearerAuth: []
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: A list of bookmarks
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    bookmark_id:
                      type: integer
                      example: 1
                    note_id:
                      type: integer
                      example: 1
                    timestamp:
                      type: string
                      example: "00:00:00"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "You are not authorized to view bookmarks for this note."
        404:
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "No bookmarks found for the note."
...
  /memos/note/{noteId}:
    get:
      tags:
        - Memos
      summary: Get memos by note ID
      description: Retrieves memos for a specific note.
      operationId: getMemosByNoteId
      security:
        - bearerAuth: []
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: A list of memos
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    memo_id:
                      type: integer
                      example: 1
                    note_id:
                      type: integer
                      example: 1
                    timestamp:
                      type: string
                      example: "00:00:00"
                    text:
                      type: string
                      example: "This is a memo"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "You are not authorized to view memos for this note."
        404:
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "No memos found for the note."
...